python-requirements-dev.txt
bs4==0.0.1
dateparser==0.7.0
gensim==3.4.0
https://github.com/LexPredict/datefinder/archive/master.zip
elasticsearch==5.4.0
ipdb==0.10.3
lxml==4.2.0
memory-profiler==0.47
nltk==3.4.5
nose==1.3.7
num2words==0.5.7
pandas==0.21.0
psutil==5.4.0
pycountry==18.5.26
pytest-cov==2.5.1
pytest-pep8==1.0.6
pytest-pylint==0.7.1
pytest-xdist==1.20.1
reporters-db==1.0.12.1
scipy==1.0.0
scikit-learn==0.21.3
sphinx==2.1.2
twine==1.13.0
Unidecode==0.4.21
us==1.0.0